test(questionnaire-e2e): added feature and tests for planning officers report
