+++ lpa-submissions-e2e-tests/cypress/integration/appealsQuestionnaireTaskList.feature
@@ -14,12 +14,13 @@ Feature: Task lists
     Given a LPA Planning Officer is reviewing their LPA Questionnaire task list
     Then The task <task> is available for selection
     Examples:
-      | task                           |
-      | "Accuracy Appellant Submission" |
-      | "Extra Conditions"             |
-      | "Other Appeals"                |
-      | "Upload Plans"                 |
-      | "Development Plan"             |
+      | task                              |
+      | "Accuracy Appellant Submission"   |
+      | "Extra Conditions"                |
+      | "Other Appeals"                   |
+      | "Upload Plans"                    |
+      | "Upload Planning Officers report" |
+      | "Development Plan"                |
 
   Scenario Outline: User is able to navigate to tasklist page by clicking on back button
     Given a LPA Planning Officer is reviewing their LPA Questionnaire task list
@@ -27,21 +28,23 @@ Feature: Task lists
     When User clicks on Back button
     Then the LPA Planning Officer is taken to the Task List
     Examples:
-      | task                           |
-      | "Accuracy Appellant Submission" |
-      | "Extra Conditions"             |
-      | "Other Appeals"                |
-      | "Upload Plans"                 |
-      | "Development Plan"             |
+      | task                              |
+      | "Accuracy Appellant Submission"   |
+      | "Extra Conditions"                |
+      | "Other Appeals"                   |
+      | "Upload Plans"                    |
+      | "Upload Planning Officers report" |
+      | "Development Plan"                |
 
   Scenario Outline: When the lpa questionnaire is not started, then the tasks are in "NOT STARTED" state
     Given a LPA Planning Officer is reviewing their LPA Questionnaire task list
     Then The state for <task> is displayed to be "NOT STARTED"
     And The state for "Before You submit - Check your answers" is displayed to be "CANNOT START YET"
     Examples:
-      | task                           |
-      | "Accuracy Appellant Submission" |
-      | "Extra Conditions"             |
-      | "Other Appeals"                |
-      | "Upload Plans"                 |
-      | "Development Plan"             |
+      | task                              |
+      | "Accuracy Appellant Submission"   |
+      | "Extra Conditions"                |
+      | "Other Appeals"                   |
+      | "Upload Plans"                    |
+      | "Upload Planning Officers report" |
+      | "Development Plan"                |
