+++ lpa-submissions-e2e-tests/cypress/integration/common/checkYourAnswers.js
@@ -1,4 +1,4 @@
-import { Given, Before } from 'cypress-cucumber-preprocessor/steps';
+import { Given, When, Then, Before } from 'cypress-cucumber-preprocessor/steps';
 import { getSubTaskInfo } from '../../support/common/subTasks';
 
 const pageId = 'confirm-answers'
@@ -13,6 +13,10 @@ Given('a change to answer {string} is requested from Change your answers page', 
   cy.goToCheckYourAnswersPage();
   currentSubTask = getSubTaskInfo(answer);
   cy.clickOnSubTaskLink(currentSubTask.id);
+});
+
+When('Check your Answers is displayed', () => {
+  cy.goToCheckYourAnswersPage();
 });
 
 Then('progress is made to the Check Your Answers page', () => {
