+++ lpa-submissions-e2e-tests/cypress/integration/common/common.js
@@ -1,5 +1,9 @@
-import { Given } from 'cypress-cucumber-preprocessor/steps';
+import { Given, When } from 'cypress-cucumber-preprocessor/steps';
 
 Given('all the mandatory questions for the questionnaire have been completed', () => {
   cy.completeQuestionnaire();
+});
+
+When('Back is then requested', () => {
+  cy.clickBackButton();
 });
