+++ lpa-submissions-e2e-tests/cypress/support/common/goToPage.js
@@ -1,6 +1,6 @@
 import defaultPathId from '../../utils/defaultPathId';
 
-module.exports = (url, id = defaultPathId) => {
-    const path = `/${id}/${url}`;
-    cy.visit(path, {failOnStatusCode:false});
- };
+module.exports = (url, id = defaultPathId, disableJs) => {
+  const path = `/${id}/${url}`;
+  cy.visit(path, { failOnStatusCode: false, script: !disableJs });
+};
